+++ dms-presentation/src/main/kotlin/team/aliens/dms/template/EmailTemplateWebAdapter.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms
+package team.aliens.dms.template
 
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import team.aliens.dms.domain.TemplateResponse
-import team.aliens.dms.domain.TemplateUseCase
+import team.aliens.dms.domain.template.usecase.TemplateResponse
+import team.aliens.dms.domain.template.usecase.TemplateUseCase
 
 @RequestMapping("/templates")
 @RestController
